Refactor Tag Autocomplete component
